Remove debugging and fix oauth bug
Prior to this commit we weren't use the @client created with the
oauth token to make api requests which meant we were hitting the
lower ratelimit of 60.

Now we're using the authorized client for every request so we have
a rate limit of 5,000 which is much easier to work with.